requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Flask-Bcrypt
tensorflow
opencv-python-headless
numpy
pymysql
python-dotenv